Mention Doxygen-style comments in style guide (#8)
We are following this convention implicitly, making it explicit. The reason for using Doxygen-style comments even for non-public APIs is that these APIs often become public at some point, but the comments are not necessarily updated to reflect that. Also, simpler rules (i.e. not thinking whether you are working on a public API or not) are simpler to follow.