Include list of docs in section pages.
Add a default layout for section pages to include the list of pages in
the section at the bottom of the section page.

For example, the 'Dialects' section page is currently empty. With this
change, the 'Dialects' page includes a list of links to all of the
dialects docs (including autogenerated dialect docs), under the heading
'Dialects Docs'.

The list of subpages is added to section pages with and without content
from a provided _index.md, unless there are no subpages in the section.