Setp site max-width to 1200px
This significantly improves readability on wide and large screens by
removing the need for the reader to scan the entire screen and
concentrating the whitespace on the edges. This is consistent with the
best practices adopted by Google search results, Github, and multiple
press websites that are text-heavy, similarly to documentation websites.