Add sphinx docs for python bindings and scripts for CI (#235)
It closes #97.

This PR adds sphinx to generate API documentation for python bindings.
It is achieved via the
[sphinx-autoapi](https://github.com/readthedocs/sphinx-autoapi) since
there are several issues while the vanilla sphinx is used for MLIR
python (mostly due to namespace package and importing).

If the workflow runs without errors (that's what I expect, although in
most cases we need some interactive rounds to debug the workflow) and
after this PR is merged, the API docs should be available in
https://mlir.llvm.org/python-bindings/.

cc @jpienaar @joker-eph @makslevental